Dev: Better upload support for .NET Core.
The main upshot here is that it is now possible for .NET to use the `__ID__` and `__NAME__` "macros" in custom strings for the database, not just the system and web paths. This gives the ability to specify the web path directly since there doesn't appear to be a reliable way to get the root web path without a dependancy on the hosting environment.