Fix: Manually added options would result in an error being thrown. Updated: New `Options.Add(Dictionary)` method which can be used to add options with the same structure as the database result and allows the options to be passed through the rendering function. Adding using the other `Options.Add()` overloads will not be passed through the rendering function.
The issue was that the added object is passed through the formatter, but
the label is already a string, not a db row, thus the error.

Don't want to just always add the options as they need to be searched
for Tags / AutoComplete. Possibly need an option in future to have added
options always appear?

https://datatables.net/forums/discussion/80574/